+++ features/support/env.rb
@@ -3,5 +3,5 @@ require 'aruba/cucumber'
 Before do
   @dirs << "../../features/fixtures"
   
-  ENV['PATH'] = "./bin#{File::PATH_SEPARATOR}#{ENV['PATH']}"  
+  ENV['PATH'] = "#{ENV['PWD']}/src#{File::PATH_SEPARATOR}#{ENV['PATH']}"  
 end