+++ src/reporting.cc
@@ -18,10 +18,12 @@
 #include "refract/TypeQueryVisitor.h"
 
 #include "refract/VisitorUtils.h"
+#include "SourceMapUtils.h"
 
 namespace sc = snowcrash;
 
 using namespace refract;
+using namespace drafter;
 
 namespace
 {
@@ -35,72 +37,6 @@ namespace
     {
         out << obj.get();
         return out;
-    }
-
-    /** structure contains starting and ending position of a error/warning. */
-    struct AnnotationPosition {
-        size_t fromLine;
-        size_t fromColumn;
-        size_t toLine;
-        size_t toColumn;
-    };
-
-    /**
-     *  \brief Convert character index mapping to line and column number
-     *  \param linesEndIndex Vector containing indexes of end line characters
-     *  \param range Character index mapping as input
-     *  \param out Position of the given range as output
-     */
-    void GetLineFromMap(const std::vector<size_t>& linesEndIndex, const mdp::Range& range, AnnotationPosition& out)
-    {
-
-        std::vector<size_t>::const_iterator annotationPositionIt;
-
-        out.fromLine = 0;
-        out.fromColumn = 0;
-        out.toLine = 0;
-        out.toColumn = 0;
-
-        // Finds starting line and column position
-        annotationPositionIt = std::upper_bound(linesEndIndex.begin(), linesEndIndex.end(), range.location) - 1;
-
-        if (annotationPositionIt != linesEndIndex.end()) {
-
-            out.fromLine = std::distance(linesEndIndex.begin(), annotationPositionIt) + 1;
-            out.fromColumn = range.location - *annotationPositionIt + 1;
-        }
-
-        // Finds ending line and column position
-        annotationPositionIt
-            = std::lower_bound(linesEndIndex.begin(), linesEndIndex.end(), range.location + range.length) - 1;
-
-        if (annotationPositionIt != linesEndIndex.end()) {
-
-            out.toLine = std::distance(linesEndIndex.begin(), annotationPositionIt) + 1;
-            out.toColumn = (range.location + range.length) - *annotationPositionIt + 1;
-
-            if (*(annotationPositionIt + 1) == (range.location + range.length)) {
-                out.toColumn--;
-            }
-        }
-    }
-
-    /**
-     *  \brief Given the source returns the length of all the lines in source as a vector
-     *  \param source Source data
-     *  \param out Vector containing indexes of all end line character in source
-     */
-    void GetLinesEndIndex(const std::string& source, std::vector<size_t>& out)
-    {
-
-        out.push_back(0);
-
-        for (size_t i = 0; i < source.length(); i++) {
-
-            if (source[i] == '\n') {
-                out.push_back(i + 1);
-            }
-        }
     }
 
     void PrintAnnotation(const std::string& prefix,
@@ -119,10 +55,10 @@ namespace
             std::cerr << " " << annotation.message;
         }
 
-        std::vector<size_t> linesEndIndex;
+        NewLinesIndex linesEndIndex;
 
         if (useLineNumbers) {
-            GetLinesEndIndex(source, linesEndIndex);
+            linesEndIndex = GetLinesEndIndex(source);
         }
 
         if (!annotation.location.empty()) {
@@ -133,8 +69,7 @@ namespace
 
                 if (useLineNumbers) {
 
-                    AnnotationPosition annotationPosition;
-                    GetLineFromMap(linesEndIndex, *it, annotationPosition);
+                    auto annotationPosition = GetLineFromMap(linesEndIndex, *it);
 
                     std::cerr << "; line " << annotationPosition.fromLine << ", column "
                               << annotationPosition.fromColumn;
@@ -158,7 +93,7 @@ namespace
         AnnotationToString(const std::string& source, const bool useLineNumbers) : useLineNumbers(useLineNumbers)
         {
             if (useLineNumbers) {
-                GetLinesEndIndex(source, linesEndIndex);
+                linesEndIndex = GetLinesEndIndex(source);
             }
         }
 
@@ -174,9 +109,8 @@ namespace
 
                     if (useLineNumbers) {
 
-                        AnnotationPosition annotationPosition;
                         mdp::Range pos(static_cast<std::int64_t>(loc->get()), static_cast<std::int64_t>(len->get()));
-                        GetLineFromMap(linesEndIndex, pos, annotationPosition);
+                        const auto annotationPosition = GetLineFromMap(linesEndIndex, pos);
 
                         output << "; line " << annotationPosition.fromLine << ", column "
                                << annotationPosition.fromColumn;